fix py2 -> py3, move deprecation warning
