+++ src/diffpy/srmise/baselines/base.py
@@ -170,8 +170,9 @@ if __name__ == "__main__":
 
     from numpy.random import randn
 
-    from diffpy.srmise.modelevaluators import AICc
-    from diffpy.srmise.peaks import GaussianOverR, Peaks
+    from diffpy.srmise.modelevaluators.aicc import AICc
+    from diffpy.srmise.peaks.base import Peaks
+    from diffpy.srmise.peaks.gaussianoverr import GaussianOverR
 
     res = 0.01
     r = np.arange(2, 4, res)
