+++ src/diffpy/srmise/baselines/fromsequence.py
@@ -145,14 +145,14 @@ class FromSequence(BaselineFunction):
             raise ValueError(emsg)
         try:
             if r[0] < self.minx or r[-1] > self.maxx:
-                logger.warn(
+                logger.warning(
                     "Warning: Evaluating interpolating function over %s, outside safe range of %s.",
                     [r[0], r[-1]],
                     [self.minx, self.maxx],
                 )
         except (IndexError, TypeError):
             if r < self.minx or r > self.maxx:
-                logger.warn(
+                logger.warning(
                     "Warning: Evaluating interpolating function at %s, outside safe range of %s.",
                     r,
                     [self.minx, self.maxx],
