+++ src/diffpy/srmise/modelevaluators/aic.py
@@ -68,7 +68,7 @@ class AIC(ModelEvaluator):
         n = fit.size
 
         if n < self.minpoints(k):
-            logger.warn("AIC.evaluate(): too few data to evaluate quality reliably.")
+            logger.warning("AIC.evaluate(): too few data to evaluate quality reliably.")
             n = self.minpoints(k)
 
         if self.chisq is None:
@@ -115,7 +115,7 @@ class AIC(ModelEvaluator):
 
         # assert n >= self.minPoints(kActual) #check that AIC is defined for the actual fit
         if n < self.minpoints(k_actual):
-            logger.warn("AIC.growth_justified(): too few data to evaluate quality reliably.")
+            logger.warning("AIC.growth_justified(): too few data to evaluate quality reliably.")
             n = self.minpoints(k_actual)
 
         penalty = self.parpenalty(k_test, n) - self.parpenalty(k_actual, n)
